+++ src/lib.rs
@@ -5,13 +5,13 @@ extern crate vgrs;
 
 use vgrs::memcheck::{make_mem_undefined, make_mem_defined};
 
-pub fn poison(addr: *const (), len: uint) {
+pub fn poison(addr: *const (), len: usize) {
     unsafe {
         make_mem_undefined(addr, len);
     }
 }
 
-pub fn unpoison(addr: *const (), len: uint) {
+pub fn unpoison(addr: *const (), len: usize) {
     unsafe {
         make_mem_defined(addr, len);
     }
